requirements.txt
matplotlib==3.9.2
numpy==2.0.1
packaging==24.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.14.1
setuptools==72.1.0
